+++ src/main/java/com/example/musictheory/services/QuizService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,30 +29,6 @@ public class QuizService {
         return Optional.ofNullable(question.orElseThrow(() -> new FileNotFoundException("Question object does not exist in database. Retry with existing uid.")));
 
     }
-
-//    public Question createQuestion(Question question) {
-//        String uid;
-//        do {
-//            uid = util.generateUid();
-//        } while (quizRepository.findQuestionByUid(uid).isPresent());
-//
-//        System.out.println("UID :: " + uid);
-//
-//        Question newQuestion = new Question();
-//        newQuestion.setUid(uid);
-//        newQuestion.setQuestion(question.getQuestion());
-//        newQuestion.setOptions(question.getOptions());
-//        newQuestion.setAnswer(question.getAnswer());
-//        if(question.getImage() != null && !question.getImage().isBlank()){
-//            newQuestion.setImage(question.getImage());
-//        }
-//        if(question.getProficiency() != null && !question.getProficiency().isBlank()){
-//            newQuestion.setProficiency(question.getProficiency());
-//        }
-//        System.out.println(newQuestion.toString());
-//
-//        return quizRepository.save(newQuestion);
-//    }
 
     public Question createQuestion(Question question) {
         String uid;
@@ -77,7 +54,7 @@ public class QuizService {
         return quizRepository.save(newQuestion);
     }
 
-    public Question updateQuestion(Question updatedQuestion) throws FileNotFoundException {
+    public Question updateQuestion(Question updatedQuestion) throws IOException {
 
         Optional<Question> question = quizRepository.findQuestionByUid(updatedQuestion.getUid());
         if (question.isPresent()) {
@@ -104,11 +81,15 @@ public class QuizService {
         }
     }
 
-    public void deleteSoftQuestion(String uid) throws FileNotFoundException {
+    public void softDeleteQuestion(String uid) throws FileNotFoundException {
 
         Optional<Question> question = quizRepository.findQuestionByUid(uid);
-        question.orElseThrow(()  -> new FileNotFoundException("Question object does not exist in database. Retry with existing uid."));
-        quizRepository.delete(question.get());
+        if (question.isPresent()) {
+            quizRepository.delete(question.get());
+        } else {
+            throw new FileNotFoundException("Uid not Found. Retry with valid uid");
+        }
+
     }
 
     public void deleteQuestion(String uid) throws FileNotFoundException {
